Refactor navigation components and layouts for improved structure
- Removed TopNav component and integrated PrimaryNav and SecondaryNav into new BaseLayout and ServicePage layouts.
- Updated PrimaryNav to eliminate unnecessary props and handle URL changes dynamically.
- Enhanced SecondaryNav to manage active sections based on the current path.
- Adjusted various pages to utilize the new layouts for better navigation consistency.
