+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from "../layouts/Skeleton.astro";
+import BaseLayout from "../layouts/Base.astro";
 import Hero from "../components/Homepage/Hero.astro";
 import UseCases from "../components/Homepage/UseCases.astro";
 import Founder from "../components/Homepage/Founder.astro";
@@ -7,7 +7,7 @@ import HowItWorks from "../components/Homepage/HowItWorks.astro";
 import Benefits from "../components/Homepage/Benefits.astro";
 ---
 
-<Layout
+<BaseLayout
   title="User- and developer-friendly authentication based on DNS"
   description="VeraId is a decentralised, user-friendly, offline-first authentication protocol powered by DNS"
 >
@@ -16,4 +16,4 @@ import Benefits from "../components/Homepage/Benefits.astro";
   <Benefits />
   <HowItWorks />
   <Founder />
-</Layout>
+</BaseLayout>
