+++ src/pages/kliento/index.astro
@@ -1,8 +1,8 @@
 ---
-import Layout from "../../layouts/Skeleton.astro";
+import ServicePage from "../../layouts/ServicePage.astro";
 ---
 
-<Layout title="Kliento" description="Kliento">
+<ServicePage title="Kliento" description="Kliento" service="Kliento">
   <section class="py-16">
     <div class="px-4 mx-auto max-w-5xl sm:px-6 lg:px-8">
       <h1 class="mb-4 text-6xl font-bold text-center leading-tight">
@@ -10,4 +10,4 @@ import Layout from "../../layouts/Skeleton.astro";
       </h1>
     </div>
   </section>
-</Layout>
+</ServicePage>
